SDIT-2778 Refactor to extract channel logic
Extract all Channel and paging logic to ReconciliationLongPagingHelper.kt file.

The helper is typed, though only supports the paging pattern that returns a last ID that will be supplied to retrieve the next page of IDs. That ID is currently hardcoded to Long given NOMIS will typically use a Long when paging via an ID is used.
